+++ src/components/common/markets/clob-widget/clob-market-trade-form.tsx
@@ -24,7 +24,7 @@ import {
 } from '@/services'
 import { useAxiosPrivateClient } from '@/services/AxiosPrivateClient'
 import useGoogleAnalytics, { GAEvents, Purchase } from '@/services/GoogleAnalytics'
-import { PendingTradeData } from '@/services/PendingTradeServise'
+import { PendingTradeData } from '@/services/PendingTradeService'
 import { useWeb3Service } from '@/services/Web3Service'
 import { paragraphMedium, paragraphRegular } from '@/styles/fonts/fonts.styles'
 import { NumberUtil } from '@/utils'
